Move Registration into its own submodule
